optimise

+++ controller/userControl.js
@@ -1,5 +1,6 @@
 const path = require("path");
 const Expense = require("../models/userExpense");
+const User = require("../models/user");
 
 exports.user = async (req, res, next) => {
   await res.sendFile(path.join(__dirname, "../views", "user.html"), (err) => {
@@ -16,12 +17,22 @@ exports.postAddExpense = async (req, res, next) => {
   try {
     const { Expenseamount, description, category } = req.body;
     const userId = req.user.id;
+    if (Expenseamount == undefined || Expenseamount.length === 0) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Parameters missing!" });
+    }
     const data = await Expense.create({
       Expenseamount: Expenseamount,
       description: description,
       category: category,
       userId: userId,
     });
+    const totalExpense = Number(req.user.totalExpenses) + Number(Expenseamount);
+    await User.update(
+      { totalExpenses: totalExpense },
+      { where: { id: req.user.id } }
+    );
     res.status(201).json(data);
     console.log(data);
   } catch (err) {
@@ -44,6 +55,13 @@ exports.postDeleteExpense = async (req, res, next) => {
   try {
     const ExpenseId = req.params.id;
     const userId = req.user.id;
+    const eAmount = await Expense.findOne({ where: { id: ExpenseId, userId } });
+    if (!eAmount) {
+      return res.status(404).json({
+        success: false,
+        message: "This Expense doesnt belongs to the user",
+      });
+    }
     const del = await Expense.destroy({ where: { id: ExpenseId, userId } });
     if (del === 0) {
       return res.status(400).json({
@@ -51,6 +69,12 @@ exports.postDeleteExpense = async (req, res, next) => {
         message: "Expense does not belong to the user.",
       });
     } else {
+      const totalExpense =
+        Number(req.user.totalExpenses) - Number(eAmount.Expenseamount);
+      await User.update(
+        { totalExpenses: totalExpense },
+        { where: { id: req.user.id } }
+      );
       res.sendStatus(200);
     }
   } catch (err) {
